requirements.txt
numpy>=1.21.0
pandas>=1.3.0
scipy>=1.7.0
scikit-learn>=1.0.0
matplotlib>=3.5.0
seaborn>=0.11.0
statsmodels>=0.13.0
pytest>=6.0.0
pytest-cov>=2.12.0
requests>=2.25.0
yfinance>=0.1.70
PyYAML>=6.0
